3577: AC.rust (#1212) - AC,100.00%,33.33%

+++ Codes/3577-count-the-number-of-computer-unlocking-permutations.rs
@@ -2,7 +2,7 @@
  * @Author: LetMeFly
  * @Date: 2025-12-10 22:34:20
  * @LastEditors: LetMeFly.xyz
- * @LastEditTime: 2025-12-10 23:04:47
+ * @LastEditTime: 2025-12-10 23:06:41
  */
 impl Solution {
     pub fn count_permutations(complexity: Vec<i32>) -> i32 {
@@ -12,7 +12,7 @@ impl Solution {
             if complexity[i] <= complexity[0] {
                 return 0;
             }
-            ans = ans * i % MOD;
+            ans = ans * (i as i64) % MOD;
         }
         ans as i32
     }
